refactor: rearrange scripts
