+++ scripts/deploy/helpers/switchboardData.ts
